Abstraction
Abstracted some things away made it simpler and easier to work with still have to figure some thing out though opacity and movement next
